pyproject.toml
[project]
name = "my-nlpproject"
version = "0.1.0"
description = ""
authors = [
    {name = "KenpachiZar7",email = "giokorompilis@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"

dependencies = [
    "numpy (<2.3)",
    "pandas (>=2.3.1,<3.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "torch (>=2.1.0,<3.0.0)",
    "transformers (>=4.55.0,<5.0.0)",
    "protobuf (>=6.31.1,<7.0.0)",
    "tiktoken (>=0.10.0,<0.11.0)",
    "sentencepiece (>=0.1.99,<0.2.0)",
    "openai (>=1.99.6,<2.0.0)",
    "sentence-transformers (>=5.1.0,<6.0.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "gensim (>=4.3.3,<5.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "fasttext-wheel (>=0.9.2,<0.10.0)"
    
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
